Add files via upload
This whole work has code for the front end and back end. To run it, you have to use a software like XAMP to host your work. You will also need to set a database and make the necessary changes on the db_credentials file to fit to your database. This work was done using the database provided by XAMP.